Modify the test module, add the file upload test, and add the image editing api
